Not working - committing my day's work of speed improvements
I also made new tokens.

Opacity filters for IE do not properly get tagged as functions.
Fonts that specify a line height (.../1.5) have the '/' not tagged as a token
Something's wrong with the block close token matching

The good news is that speed is now sliced to 1/40th of what it was for big
files.